feat: rename and drop view
