+++ dms-plugins/dms-plugins-datasource/src/main/java/com/basedt/dms/plugins/datasource/impl/clickhouse/ClickHouseFgnTableHandler.java
@@ -1,0 +1,35 @@
+package com.basedt.dms.plugins.datasource.impl.clickhouse;
+
+import cn.hutool.core.util.StrUtil;
+import com.basedt.dms.plugins.datasource.dto.TableDTO;
+import com.basedt.dms.plugins.datasource.impl.jdbc.JdbcForeignTableHandler;
+
+import java.sql.SQLException;
+import java.util.List;
+
+public class ClickHouseFgnTableHandler extends JdbcForeignTableHandler {
+
+    @Override
+    public List<TableDTO> listForeignTables(String catalog, String schemaPattern, String tablePattern) throws SQLException {
+        String sql = " select" +
+                "    t.table_catalog as catalog_name," +
+                "    t.table_schema as schema_name," +
+                "    t.table_name as object_name," +
+                "    'FOREIGN_TABLE' as object_type," +
+                "    t.table_comment as remark," +
+                "    st.metadata_modification_time as create_time," +
+                "    st.metadata_modification_time as last_ddl_time" +
+                " from information_schema.tables t" +
+                " left join system.tables st" +
+                " on t.table_schema = st.database" +
+                " and t.table_name = st.name" +
+                " where t.table_type in ('FOREIGN TABLE')";
+        if (StrUtil.isNotEmpty(schemaPattern)) {
+            sql += " and t.table_schema = '" + schemaPattern + "'";
+        }
+        if (StrUtil.isNotEmpty(tablePattern)) {
+            sql += " and t.table_name = '" + tablePattern + "'";
+        }
+        return super.listFgnTableFromDB(sql);
+    }
+}
